Manejo de error 500 y 404

+++ app/Services/RestauranteService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Exceptions\RestauranteNoEncontradoException;
 use App\Repositories\RestauranteRepository;
 
 class RestauranteService{
@@ -16,16 +17,26 @@ class RestauranteService{
     }
 
     public function mostrar($id){
-        return $this->repository->find($id);
+        $restaurante =  $this->repository->find($id);
+        if(!$restaurante){
+            throw new RestauranteNoEncontradoException();
+        }
+        return $restaurante;
     }
 
     public function actualizar($id,$dto){
         $restaurante = $this->repository->find($id);
+        if(!$restaurante){
+            throw new RestauranteNoEncontradoException();
+        }
         return $this->repository->update($restaurante,$dto->toArray());
     }
 
     public function eliminar($id){
         $restaurante = $this->repository->find($id);
+        if(!$restaurante){
+            throw new RestauranteNoEncontradoException();
+        }
         return $this->repository->delete($restaurante);
     }
